e2e/requirements.txt
robotframework==6.1.1
argcomplete==1.12.3
directio==1.3
flake8
kubernetes==27.2.0
requests==2.32.5
boto3==1.42.34
pyyaml==6.0.3
minio==5.0.10
robotframework-retryfailed
prometheus_client
